Moved implementation to "tlsconfig.go" also implemented suggestions. Suggestions implemented: https://github.com/roadrunner-server/redis/pull/104#discussion_r1848927383 https://github.com/roadrunner-server/redis/pull/104#discussion_r1848932114 https://github.com/roadrunner-server/redis/pull/104#discussion_r1848932950 https://github.com/roadrunner-server/redis/pull/104#discussion_r1848934883
